refactor(server): move connector machinery to connector.go, decouple resolver from Server
Extract the connector registry (ConnectorsConfig, openConnector), the local
password DB, and the resolver into server/connector.go. resolveConnector was a
Server method reading s.storage/s.logger; it becomes connectorResolver(storage,
logger) returning the connectors.ResolveFunc the cache already injects, so
connector construction no longer depends on the Server. server.go sheds all the
connector/* imports along with it.
